+++ sparkle/playground/src/components/InputBar.tsx
@@ -1,34 +1,59 @@
 import {
   ArrowUpIcon,
   AttachmentIcon,
+  Avatar,
   BoltIcon,
   Button,
   cn,
+  Dialog,
+  DialogContainer,
+  DialogContent,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
   DocumentIcon,
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuSearchbar,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
   Icon,
   ImageIcon,
   ImageZoomDialog,
   MicIcon,
   PlusIcon,
-  RobotIcon,
   Sheet,
   SheetContainer,
   SheetContent,
   SheetHeader,
   SheetTitle,
 } from "@dust-tt/sparkle";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+
+import { getAgentAvatarProps, getAgentById, mockAgents } from "../data";
 
 import { NewCitation, NewCitationGrid } from "./NewCitation";
 import { RichTextArea, type RichTextAreaHandle } from "./RichTextArea";
 
 type DroppedFile = { id: string; file: File; objectUrl?: string };
 
+export interface InputBarForkWithContext {
+  sourceConversationId: string;
+  onConfirm: (newAgentId: string) => void;
+}
+
 interface InputBarProps {
   placeholder?: string;
   className?: string;
   instructionReference?: { start: number; end: number } | null;
   onInstructionInserted?: () => void;
+  /** When set with initialAgentId, switching threads resets the bar agent. */
+  conversationKey?: string;
+  /** Primary agent for the current thread (e.g. first agent participant). */
+  initialAgentId?: string;
+  /** When set, choosing another agent opens a confirm dialog before forking. */
+  forkWithContext?: InputBarForkWithContext;
 }
 
 export function InputBar({
@@ -36,6 +61,9 @@ export function InputBar({
   className,
   instructionReference,
   onInstructionInserted,
+  conversationKey,
+  initialAgentId,
+  forkWithContext,
 }: InputBarProps) {
   const [isFocused, setIsFocused] = useState(false);
   const [isDragOver, setIsDragOver] = useState(false);
@@ -44,6 +72,14 @@ export function InputBar({
     useState<DroppedFile | null>(null);
   const [isImageZoomOpen, setIsImageZoomOpen] = useState(false);
   const [isCitationSheetOpen, setIsCitationSheetOpen] = useState(false);
+  const [mentionAgentSearch, setMentionAgentSearch] = useState("");
+  const [selectedAgentId, setSelectedAgentId] = useState(
+    initialAgentId && initialAgentId !== "" ? initialAgentId : "agent-dust"
+  );
+  const [isForkDialogOpen, setIsForkDialogOpen] = useState(false);
+  const [pendingForkAgentId, setPendingForkAgentId] = useState<string | null>(
+    null
+  );
   const containerRef = useRef<HTMLDivElement>(null);
   const richTextAreaRef = useRef<RichTextAreaHandle | null>(null);
   const dragCounterRef = useRef(0);
@@ -154,6 +190,39 @@ export function InputBar({
     onInstructionInserted?.();
   }, [instructionReference, onInstructionInserted]);
 
+  useEffect(() => {
+    if (initialAgentId != null && initialAgentId !== "") {
+      setSelectedAgentId(initialAgentId);
+    }
+  }, [conversationKey, initialAgentId]);
+
+  const selectedAgent = useMemo(
+    () => getAgentById(selectedAgentId) ?? mockAgents[0],
+    [selectedAgentId]
+  );
+
+  const agentTriggerVisual = useMemo(
+    () =>
+      function AgentTriggerVisual({ className }: { className?: string }) {
+        return (
+          <Avatar
+            className={className}
+            size="xxs"
+            {...getAgentAvatarProps(selectedAgent)}
+          />
+        );
+      },
+    [selectedAgent]
+  );
+
+  const filteredMentionAgents = useMemo(() => {
+    const q = mentionAgentSearch.trim().toLowerCase();
+    const base = q
+      ? mockAgents.filter((a) => a.name.toLowerCase().includes(q))
+      : mockAgents;
+    return [...base].sort((a, b) => a.name.localeCompare(b.name));
+  }, [mentionAgentSearch]);
+
   const showFocusStyle = isFocused || isDragOver;
 
   return (
@@ -212,7 +281,66 @@ export function InputBar({
           showAskSidekickMenu={false}
           className="placeholder:s-text-muted-foreground dark:placeholder:s-text-muted-foreground-night"
         />
-        <div className="s-flex s-w-full s-gap-2 s-p-2 s-pl-4">
+        <div className="s-flex s-w-full s-gap-2 s-p-2.5">
+          <DropdownMenu
+            onOpenChange={(open) => {
+              if (!open) setMentionAgentSearch("");
+            }}
+            modal={false}
+          >
+            <DropdownMenuTrigger asChild>
+              <Button
+                variant="ghost-secondary"
+                icon={agentTriggerVisual}
+                size="sm"
+                label={selectedAgent.name}
+                tooltip="Mention an agent"
+                isSelect
+                className="s-max-w-[44vw] sm:s-max-w-[11rem]"
+                onClick={(e: React.MouseEvent) => {
+                  e.stopPropagation();
+                }}
+              />
+            </DropdownMenuTrigger>
+            <DropdownMenuContent
+              align="start"
+              className="s-w-[min(380px,calc(100vw-2rem))]"
+              dropdownHeaders={
+                <DropdownMenuSearchbar
+                  value={mentionAgentSearch}
+                  onChange={setMentionAgentSearch}
+                  name="input-bar-mention-agent"
+                  placeholder="Search agents"
+                />
+              }
+            >
+              <DropdownMenuSeparator />
+              {filteredMentionAgents.length > 0 ? (
+                filteredMentionAgents.map((agent) => (
+                  <DropdownMenuItem
+                    key={agent.id}
+                    label={agent.name}
+                    description={agent.description}
+                    icon={<Avatar size="sm" {...getAgentAvatarProps(agent)} />}
+                    onSelect={() => {
+                      setMentionAgentSearch("");
+                      if (forkWithContext && agent.id !== selectedAgentId) {
+                        setPendingForkAgentId(agent.id);
+                        setIsForkDialogOpen(true);
+                        return;
+                      }
+                      setSelectedAgentId(agent.id);
+                    }}
+                    truncateText
+                  />
+                ))
+              ) : (
+                <div className="s-flex s-h-24 s-items-center s-justify-center s-text-sm s-text-muted-foreground dark:s-text-muted-foreground-night">
+                  No agents found
+                </div>
+              )}
+            </DropdownMenuContent>
+          </DropdownMenu>
           <Button
             variant="outline"
             icon={PlusIcon}
@@ -224,40 +352,76 @@ export function InputBar({
             <Button
               variant="ghost-secondary"
               icon={AttachmentIcon}
-              size="xs"
+              size="sm"
               tooltip="Attach a document"
             />
             <Button
               variant="ghost-secondary"
               icon={BoltIcon}
-              size="xs"
+              size="sm"
               tooltip="Add functionality"
-            />
-            <Button
-              variant="ghost-secondary"
-              icon={RobotIcon}
-              size="xs"
-              tooltip="Mention an Agent"
             />
           </div>
           <div className="s-grow" />
           <div className="s-flex s-items-center s-gap-2 md:s-gap-1">
-            <Button
-              variant="ghost-secondary"
-              icon={MicIcon}
-              size="xs"
-              isRounded
-            />
+            <Button variant="ghost-secondary" icon={MicIcon} size="sm" />
             <Button
               variant="highlight"
               icon={ArrowUpIcon}
-              size="xs"
+              size="sm"
               tooltip="Send message"
               isRounded
             />
           </div>
         </div>
       </div>
+
+      <Dialog
+        open={isForkDialogOpen}
+        onOpenChange={(open) => {
+          setIsForkDialogOpen(open);
+          if (!open) setPendingForkAgentId(null);
+        }}
+      >
+        <DialogContent size="md">
+          <DialogHeader>
+            <DialogTitle>New conversation</DialogTitle>
+          </DialogHeader>
+          <DialogContainer>
+            <p className="s-text-sm s-text-foreground dark:s-text-foreground-night">
+              Would you like to open a new conversation with{" "}
+              <span className="s-font-semibold">
+                {pendingForkAgentId
+                  ? (getAgentById(pendingForkAgentId)?.name ?? "this agent")
+                  : ""}
+              </span>{" "}
+              and bring this conversation as context?
+            </p>
+          </DialogContainer>
+          <DialogFooter
+            leftButtonProps={{
+              label: "Cancel",
+              variant: "outline",
+              onClick: () => {
+                setIsForkDialogOpen(false);
+                setPendingForkAgentId(null);
+              },
+            }}
+            rightButtonProps={{
+              label: "New conversation",
+              variant: "highlight",
+              onClick: () => {
+                if (pendingForkAgentId && forkWithContext) {
+                  forkWithContext.onConfirm(pendingForkAgentId);
+                  setSelectedAgentId(pendingForkAgentId);
+                }
+                setIsForkDialogOpen(false);
+                setPendingForkAgentId(null);
+              },
+            }}
+          />
+        </DialogContent>
+      </Dialog>
 
       {/* Image preview dialog */}
       {selectedDroppedFile?.objectUrl && (
