+++ sparkle/playground/src/data/index.ts
@@ -2,6 +2,7 @@
 export * from "./types";
 
 // Re-export all data and utilities
+export * from "./agentAvatarProps";
 export * from "./agentBuilder";
 export * from "./agents";
 export * from "./conversations";
