+++ sparkle/playground/src/stories/Projects.tsx
@@ -77,12 +77,14 @@ import {
   getRandomSpaces,
   getRandomUsers,
   getUserById,
+  getAgentAvatarProps,
   mockAgents,
   mockConversations,
   mockUsers,
   type Space,
   type User,
 } from "../data";
+import { createForkedConversation } from "../utils/createForkedConversation";
 
 type Collaborator =
   | { type: "agent"; data: Agent }
@@ -453,9 +455,7 @@ function DustMain() {
                         ) : (
                           <Avatar
                             size="xxs"
-                            name={participant.data.name}
-                            emoji={participant.data.emoji}
-                            backgroundColor={participant.data.backgroundColor}
+                            {...getAgentAvatarProps(participant.data)}
                             isRounded={false}
                           />
                         )
@@ -1180,6 +1180,24 @@ function DustMain() {
     // For prototype, just log the update
   };
 
+  const handleForkConversationWithContext = ({
+    newAgentId,
+    sourceConversation,
+  }: {
+    newAgentId: string;
+    sourceConversation: Conversation;
+  }) => {
+    if (!user) return;
+    const newConv = createForkedConversation({
+      source: sourceConversation,
+      newAgentId,
+      locutorUserId: user.id,
+    });
+    setConversationsWithMessages((prev) => [...prev, newConv]);
+    setSelectedConversationId(newConv.id);
+    setSelectedView(null);
+  };
+
   // Main content
   const mainContent =
     // Priority 0: Show profile when opened from user menu
@@ -1206,6 +1224,7 @@ function DustMain() {
         showBackButton={!!previousSpaceId}
         onBack={handleConversationBack}
         projectTitle={selectedSpace?.name}
+        onForkConversationWithContext={handleForkConversationWithContext}
       />
     ) : // Priority 3: Show space view if a space is selected
     selectedSpace && selectedSpaceId ? (
